// src/api/generated/models/Holiday.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Blackwell Time API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Holiday
 */
export interface Holiday {
    /**
     * 
     * @type {string}
     * @memberof Holiday
     */
    holidayId?: string;
    /**
     * 
     * @type {string}
     * @memberof Holiday
     */
    holidayName?: string;
    /**
     * 
     * @type {Date}
     * @memberof Holiday
     */
    holidayDate?: Date;
}

/**
 * Check if a given object implements the Holiday interface.
 */
export function instanceOfHoliday(value: object): value is Holiday {
    return true;
}

export function HolidayFromJSON(json: any): Holiday {
    return HolidayFromJSONTyped(json, false);
}

export function HolidayFromJSONTyped(json: any, ignoreDiscriminator: boolean): Holiday {
    if (json == null) {
        return json;
    }
    return {
        
        'holidayId': json['holidayId'] == null ? undefined : json['holidayId'],
        'holidayName': json['holidayName'] == null ? undefined : json['holidayName'],
        'holidayDate': json['holidayDate'] == null ? undefined : (new Date(json['holidayDate'])),
    };
}

export function HolidayToJSON(json: any): Holiday {
    return HolidayToJSONTyped(json, false);
}

export function HolidayToJSONTyped(value?: Holiday | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'holidayId': value['holidayId'],
        'holidayName': value['holidayName'],
        'holidayDate': value['holidayDate'] == null ? value['holidayDate'] : value['holidayDate'].toISOString().substring(0,10),
    };
}